requirements.txt
diff-match-patch==20121119
Django==2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-import-export==1.1.0
django-taggit==0.22.2
django-widget-tweaks==1.4.1
et-xmlfile==1.0.1
jdcal==1.4
Markdown==2.6.11
odfpy==1.3.6
openpyxl==2.5.9
pytz==2017.3
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
